requirements.txt
fastapi
uvicorn
pymongo
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart